vision_service/requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
pydantic
